Refactor agent util into resolve_model_call_path (#1999)
