+++ nemo_gym/base_responses_api_model.py
@@ -37,7 +37,7 @@ import threading
 import time
 from abc import abstractmethod
 from pathlib import Path
-from typing import Any, Optional
+from typing import Any, Mapping, Optional
 
 import orjson
 from fastapi import Body, FastAPI, Request
@@ -45,6 +45,12 @@ from fastapi.responses import StreamingResponse
 from pydantic import BaseModel, Field, model_validator
 
 from nemo_gym.anthropic_converter import AnthropicConverter
+from nemo_gym.config_types import ROLLOUT_PATH_PREFIX
+from nemo_gym.global_config import (
+    ATTEMPT_INDEX_KEY_NAME,
+    ROLLOUT_INDEX_KEY_NAME,
+    TASK_INDEX_KEY_NAME,
+)
 from nemo_gym.openai_utils import (
     NeMoGymChatCompletion,
     NeMoGymChatCompletionCreateParamsNonStreaming,
@@ -52,11 +58,9 @@ from nemo_gym.openai_utils import (
     NeMoGymResponseCreateParamsNonStreaming,
 )
 from nemo_gym.server_utils import (
-    ROLLOUT_PATH_PREFIX,
     BaseRunServerInstanceConfig,
     BaseServer,
     SimpleServer,
-    maybe_rollout_id_from_run_body,
 )
 
 
@@ -215,6 +219,34 @@ class CaptureStore:
                 finally:
                     fcntl.flock(handle.fileno(), fcntl.LOCK_UN)
         return exchanges
+
+
+def maybe_rollout_id_from_run_body(body: BaseModel | Mapping[str, Any] | None) -> Optional[str]:
+    """Per-rollout model-call capture id from a run-request's task/rollout indices.
+
+    Reads the canonical row keys (``_ng_task_index`` / ``_ng_rollout_index``) that
+    rollout_collection ships to an agent's ``/run``. When a resume re-dispatch attempt is present
+    (``_ng_attempt_index`` > 0), an ``-a<n>`` suffix is appended so a retry's captured model calls
+    stay separable from the prior attempt; the first attempt (0) keeps the bare ``<task>-<rollout>``
+    key for backward compatibility.
+    """
+    if isinstance(body, BaseModel):
+        data = body.model_dump()
+    elif isinstance(body, Mapping):
+        data = body
+    else:
+        return None
+    task = data.get(TASK_INDEX_KEY_NAME)
+    rollout = data.get(ROLLOUT_INDEX_KEY_NAME)
+    if task is None or rollout is None:
+        return None
+    rollout_id = f"{task}-{rollout}"
+    attempt = data.get(ATTEMPT_INDEX_KEY_NAME)
+    if attempt is not None:
+        attempt_index = int(attempt)
+        if attempt_index > 0:
+            rollout_id = f"{rollout_id}-a{attempt_index}"
+    return rollout_id
 
 
 # --- Observability records derived from captured exchanges ---
