+++ nemo_gym/config_types.py
@@ -751,7 +751,7 @@ class WANDBConfig(BaseModel):
 
 
 ########################################
-# Weights and Biases
+# Aggregate Metrics
 ########################################
 
 
@@ -784,3 +784,12 @@ class AggregateMetrics(BaseModel):
         default_factory=dict,
         description="Headline metrics for this benchmark. Subset of agent_metrics.",
     )
+
+
+########################################
+# Model Call Capture
+########################################
+
+# Per-rollout model-call correlation. Callers place the rollout id in the model-server URL;
+# the capture middleware in base_responses_api_model.py strips this prefix before routing.
+ROLLOUT_PATH_PREFIX = "ng-rollout"
